Make the cards check if they are eachothers match
This was acomplished by having a counter that checked the number of cards flipped,
if it was 2 then it forbid the user to flip any more cards and started checking
whether they were eachothers match or not, if they were we left them flipped but
disabled the option to flip them again, like in real life memorama.
If they were not a match then they were flipped back again after a short time
so that the user could see what the cards contained.