Winning configuration
In this commit all the options regarding winning the game were adressed,
starting with the internal counter to know when the user had won. After that
a winning screen was implemented and animated to drop once the game was won.
In this winning screen two buttons were also configured as options to the
user, one being the play again, which repeats the game with the same
specifications the user entered and the end game button which works as the
mid game end button.